[FIX] l10n_mx_hr_payroll: compute paid amount correctly
Issue:
======
`paid_amount` is not calculated correctly when the schedule pay isn't
month.

Steps to reproduce the issue:
=============================
- Switch to MX company
- Go to roque and employee and add these values to his contract : salary
  structure type( Mexico:Employee) - both schedule pay fields (weekly) -
  add any amount in wage.
- Go to payslips and create one for roque and click on compute sheet
- Go to salary computation tab
- Basic salary is wrong it should be the amount you put at the start.

Origin of the issue and solution:
=================================
We were dividing by the coefficient of the pay schedule but in reality
we didn't have to since we create the payslip for the period of the
schedule pay so it will have the full mentioned wage.

But for the christmas we need to divide by the coefficient since it's
computed by `days/year`.

opw-413332

closes odoo/enterprise#71161

X-original-commit: cda9c0d741118828be0fb912a49d26d63b6b429c
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>