[FIX] timesheet_grid: fix lock date error message with user's date format
Before this commit:
The lock date error message was hard-coded to the MM/DD/YYYY format, regardless
of the user's language, which causing inconsistency and confusion in date format.

After this commit:
The lock date error message now dynamically follows the date format set to the
user's language, which improve consistency and reduce confusion.

task-3925185

closes odoo/enterprise#72640

X-original-commit: a1d3368d26e4060e37d9a8676d938941baefadae
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>