[REM] sale_subscription: remove automation rules
This commits removes the dependency of 'base_automation' in the Subscription
application and its derived code. The objective of this task is removing
the code since it does not add much value and the customer can install
automations from others modules if needed (such as Studio).

task-4181465

closes odoo/enterprise#70180

Related: odoo/upgrade#6527
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>