[FIX] sale_stock_renting: allow to disable the rental transfers setting
Steps to reproduce:

- In the setting: enable Rental transfers
- Create and confirm a rental order for 1 unit of a product.
> 2 transfers are created: one delivery and one receipt
- Validate the delivery and then the receipt for your product.
- Go back to the setting: disable Rental transfers

Expected behavior:

You are expected to be able to change this setting. In addition, you
have already ended each of your flows that was sensitive to this
settings so that nothing should be blocking.

Current behavior:

Error: caused by a foreign key violation:
"stock_move_rule_id_fkey" on the table "stock_move".

Cause fo the issue:

Push and pull Rules were refactored in saas-17.2 see commit a2ea658
along with this refactoring a method was added to delete the rules of
the rental route in oreder to update them if the route is disabled:
https://github.com/odoo/enterprise/blob/4647c62c3bd05b7346033cfa851770dfc399837e/sale_stock_renting/models/stock_warehouse.py#L15-L20
However, if the rule was already used on a stock move (just like in our
flow) these rules will still be referenced by a move and you will not be
allowed to delete the rules form the "stock_rule" table without
violating the foreign key violation "stock_move_rule_id_fkey" on the
"stock_move" table.

opw-4131426

closes odoo/enterprise#70786

X-original-commit: 33d4048b607956dfdefb09bca728a65e8dbd540f
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>