[IMP] spreadsheet_dashboard: introduce sample dashboards
Currently, users with an fresh empty database will only see empty
dashboards. They will go through a list of empty tables and charts without any
data and probably miss the potential of the dashboard app.

For a proper onboarding experience, they should be greeted with a sample
dashboard to give them a better idea of the dashboard look and use.

Technically, we store sample json files to be used when the
main model (the model that is used the most in the dashboard) has no records.

task-3947773

X-original-commit: 1c97c1ea79b3913e4b7ab6ba00e3e92053808120
Part-of: odoo/enterprise#70982
Related: odoo/odoo#181962
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Mehdi Rachico (mera) <mera@odoo.com>