[FIX] l10n_pe_edi: disable tax check on combos
Currently, some localizations are not able to invoice if a combo product
was bought.

Steps to reproduce:
-------------------
* Install **l10n_pe_edi_pos**
* Switch to the **PE Company**
* Go to the **Point of sale** App
* Go to the products
* Select a combo product
* Navigate through all the products that can be selected in the combo and
apply a tax on each
* Open shop session
* Add the combo product to the order
* Validate and invoice the order
> Observation: Invalid Operation: Taxes need to be assigned on all invoice
lines

Why the fix:
------------
Combo products do not have the possibility to be assigned a tax, as they are
not supposed to. Taxes are computed for each product chosen and applies on
that product.

We can treat the parent combo line as the note or sections we would put on
a quotation.

opw-4090946

closes odoo/enterprise#71601

X-original-commit: 43d0ca746f31e8fd19fa21a5da6439a66cd9248a
Related: odoo/odoo#183037
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>