[IMP] stock_enteprise: extend map view to allow multi records summary
This commit enables the popup marker of map view to show the summary of
multiple records in `stock_enterprise` module. The standard behavior of
the map view is to show the summary on the popup marker in case it's
related to only one record. Otherwise, it only shows the address. The
extension introduced by this commit is only used by the map view of
stock pickings (which is the only map view in `stock_enterprise`),
however, this feature can be applied to any map view by using the
`js_class` defined in this commit.

closes odoo/enterprise#69747

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>