[IMP] social_twitter: update stream posts according to user_likes
This commit now enables the posts linked to the same `twitter_tweet_id` to have
the same `twitter_likes_count` when the user likes a post via the feed.

Beforehand when a post was present inside 2 separate streams it wasn't possible
for both to have the same data inside the feed.
This is caused by the fact that they are distinct records in the DB.

To fix this, when the user likes a tweet on the feed kanban view we propagate
the new values to the record objects linked to posts with the same
`twitter_tweet_id`.

task-3986158

Part-of: odoo/enterprise#64647
Related: odoo/upgrade#6158
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>