[IMP] l10n_mx: print payment receipt
This commit add a new print button on the cfdi document view for payment, which allow the user to print the payment receipt directly from the invoice.

Also, this print button allow the user to print a payment receipt from a bank transaction, as it was not possible before this commit.

task-4161205

closes odoo/enterprise#70318

Signed-off-by: Laurent Smet (las) <las@odoo.com>