[IMP] crm_helpdek: add test case for converting ticket to opportunity or lead
This commit introduces a new test case that ensures the conversion of helpdesk
tickets to opportunities or lead in the CRM module is functioning as expected.

Test Details: A new user is created, and conversion is attempted both before and
after assigning the user to the relevant group (crm.group_use_lead). Assertions
check for both "Convert to Opportunity" and "Convert to Lead" behavior depending
on the user's group permissions.

closes odoo/enterprise#71957

Task: 4180279
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>