[IMP] appointment: user capacity and multiple bookings
This commit introduces two new features for the appointment:
1. Support User Capacity
2. Manage Multiple Bookings for user and resource

1) Support User Capacity
====================
Similar to what we do with resources, we will have users to be booked for
multiple seats.

Use Case
-------------
There is a Yoga Class having two instructors (Tom and Jessica)
User capacity is set to 5.
1. Each instructor can give lesson to 5 people. (possibly spread over different
bookings)
2. The instructor do not behave like shared resources (Tom cannot be booked for
10 persons in one go using both instructors (in auto assign))

2) Multiple Bookings Per Slot
======================
This method is a new capacity management method.
From this commit we will have max number of bookings for the appointment
when the appointment doesn't manage the capacity.
The working of this method is simple. For any given available slot, a user or a
resource can be booked for the given max_bookings at max.

Use Case
------------
Max Bookings per slot is set to 4.
There is a real estate agency (Micheal and David are agents)
Micheal and David can be booked 4 times on each slot.
When you book a house visit with a real estate agent, they don't care
if you come with someone to help you with the visit. But they know they'll
accept up to 4 "potential buyers".

Task-4144546

closes odoo/enterprise#70883

Related: odoo/odoo#202079
Related: odoo/upgrade#6575
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>