[IMP] hr_payroll: auto recompute sheet on payslips
There is a very high chance of committing mistake when the payslip is
confirmed without re-computing (manually) after changing the Structure/Period
or any other field which has impact on computation. This happens especially
when dealing with batch payments, where payslips have to be re-computed
and confirmed when sent back to the draft stage to change any of the above
mentioned fields (like Structure, Period etc.)

To address this, auto recomputation is required. This change makes the
payslip recompute automatically whenever it is saved.

task-4058455

closes odoo/enterprise#68479

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>