[FIX] sale_subscription: resolve access token conflict
Versions
--------
- saas-17.2+

Steps
-----
1. Activiate Stripe, Sips, or Demo as payment provider;
2. create a subscription & confirm;
3. invoice subscription & confirm;
4. click share link & copy;
5. open link in a private window/different browser;
6. pay invoice.

Issue
-----
The access token is invalid, which is important if you as user don't
have regular access to the document (or are logged out).

Cause
-----
When a subscription payment flow is followed via an invoice,
it sets the `access_token` in the controller and adds a
`transaction_route_subscription` value.  The `access_token` is valid for
the `transaction_route_subscription`, but this route is only followed
when payments are automated (added in 36b238b934f).

Initially, the invoice token was passed, which worked for non-automated
payments getting passed as regular invoices to `account_payment`.

Commit 60fd12a6c65 changed this to a subscription token, to allow
automated payments, but broke non-automated payments in the process, as
those still go via `account_payment`, which expects an invoice token.

Solution
--------
Force all to go via invoice route using invoice token.

Also fix bug in invoice route where payment token wouldn't be assigned
if payment happenend with a saved token.

opw-4152257

closes odoo/enterprise#71316

X-original-commit: 189d5ac66cb8efdb93b1a76f61b17d12b7cd47c3
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>