[REF] l10n_mx_edi: Many2many addenda, support Complementos
This commit refactors the usage of `l10n_mx_edi.addenda` to allow adding
multiple addendas at once, and also to support injecting to Complemento
node on top of the Addenda node. On top of it, we will now also be able
to inject before sending (for complementos addenda) to make sure the
injected XML gets sent to the government.

- l10n_mx_edi_addenda_id on res_partner and account_move is now
  l10n_mx_edi_addenda_ids (Many2many)
- custom addenda xml nodes in the addenda's arch to support specifying
  where to inject the XML, and when to inject so that it may be sent to
  government
- constrains on the addenda/complemento XML arch will be done with the
  custom namespaces to allow usage of the cfdi XML elements inside them

The custom xml nodes are decided to be:

- `<t t-xml-node="(xml_node)"> ... </t>` -> for specifying where should
  we process the information contained inside it
- `<t t-namespace-key="(nskey)" t-namespace-url="(nsurl)"/>` -> for
  specifying the namespaces that we should inject to (xml_node)
- `<t t-schema-locations="(schema_locations)"/>` -> for specifying what
  schema locations to put inside the Comprobante node. This node should
  also only appear inside the `<t t-xml-node="comprobante">...</t>`
  parent element

The default arch are also updated to better guide the user on the
accepted format of the addenda arch, due to the strict nature of the
custom nodes locations/contents

closes odoo/enterprise#70569

Task-id: 3894067
Related: odoo/upgrade#6862
Signed-off-by: Laurent Smet (las) <las@odoo.com>