[FIX] whatsapp: allow public to send themselves messages
Currently it's not possible for a public user to send themselves a message
over whatsapp when the message contains information linked to a record

This is because most records do not allow public users from reading them.

As whatsapp can only used a select list of "safe" fields we can render the fields
as SU when using the cron.

task-4199766

X-original-commit: 67dc25fe2dd3031736cfd7417eb29563e1528f95
Part-of: odoo/enterprise#81737
Related: odoo/odoo#202459
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>