[REF] knowledge, website_knowledge: start using html_editor for knowledge
This commit prepares the conversion to use the new html_editor in Knowledge.
web_editor related assets are neutralized in knowledge (see remove commit for
their complete removal).

website_knowledge is refactored to stop using a Public Widget, and will now use
the HtmlViewer to display the content of a Knowledge article. This will allow to
mount embedded components in the public view.

The article plugin is adapted to handle the features of the now obsolete
article_behavior (open an article using the action service).

task-3672747

X-original-commit: 08d84f8b61450a0ebd0d2aadfe8f227bda283edd
Part-of: odoo/enterprise#71082
Related: odoo/odoo#182145
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>