[IMP] knowledge: migration to obtain embedded components from behaviors
Upgrading html values are an expensive task, since the html tree has to be
built for every field value. This commit introduces a way to modify an existing
html value coming from an earlier version of Odoo using Behaviors and the old
knowledge comments anchors to their new form.

This JS "upgrade" will be attempted once, and will maintain the existing raw
value if it fails without crashing. If it succeeds, the html elements will be
converted to their new form, and if the html value is saved afterwards, will
never have to run again. If the html value is consulted without write access, it
will be converted each time it is loaded, but the operation is pretty
inexpensive (fully synchronous, no rpc needed).

The files introduced with this commit can be removed the moment a standard
upgrade is done.

task-3672747

X-original-commit: 0df669159aaa1a9631ab12a68758bf78954071f7
Part-of: odoo/enterprise#71082
Related: odoo/odoo#182145
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>