[REF] l10n_mx_edi: rewrite getattr as map function
This commit replaces the usage of getattr in l10n_mx_edi document to a
function that returns a dictionary map of the needed PAC functions.

Purpose: getattr usage is messy. It's hard to search where the PAC
method is used in the code, and it's also a security risk. If we want to
add new PAC method, we should explicitly add it in the new pac method
map getter function so that it is clear that those new function will be
used and have a clear reference.

This will also allow us to better rename the PAC methods if needed in
the future.

Task-id: 3894067
Part-of: odoo/enterprise#70569
Related: odoo/upgrade#6862
Signed-off-by: Laurent Smet (las) <las@odoo.com>