[FIX] l10n_ec_edi_pos: make partner selectable
Before this commit, changing the partner in POS for the EC localization
was raising an OWL error.
This error happens because missingFields got stripped from most of the
codebase, including PartnerList, leading to an "Invalid Props" error in Owl.
This was preventing users from selecting a partner other than the default
"consumidor final".

After this commit, clicking the partner button correctly opens the
"partner selection" popup.

opw-4171779
opw-4192162

closes odoo/enterprise#70745

X-original-commit: 7435214132885dc3ebca8d73b68394465d915814
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>