[FIX] sale_timesheet_enterprise : Show total hours of validated timesheets only
Steps to reproduce:
	- Set timesheets inovicing policy to 'Validated Timesheets Only'
	- From the portal go to any task that has some validated timesheet and some other not
	- Notice that the lines shown is the validated ones only but the Hours spent includes
	the non-validated ones too

Current behavior before PR:
The total hours spent for timesheets in portal view when setting the
invoicing policy to 'Validate Timesheets Only' is calculated wrongly
as it takes the non validated ones into account.

This is happening as when passing the values to the view we pass 'task_sudo'
variable which will have the SUPERUSER as the env.user

https://github.com/odoo/odoo/blob/saas-17.2/addons/project/controllers/portal.py#L569

so it won't pass the condition if the user is portal or not
https://github.com/odoo/enterprise/blob/saas-17.2/sale_timesheet_enterprise/models/project_task.py#L33:L35
and it will calculate all the timesheet without taking validation into
condisderation.

Desired behavior after PR is merged:
As we are using those computed fields only in portal views so we don't
have to check if the user is portal user as even if an internal user is
showing the same view for testing he should be able to see the same data
as the portal user.

opw-4193842

closes odoo/enterprise#92301

X-original-commit: a971f55e736f9645eda137fd6dcb574483886483
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>