[FIX] web_grid: make sure to apply style only in grid view
Before this commit, the 3c7775f4c7a3efaa891bc31072c41abff24edfad commit
adds some style to correctly display the noContent helper inside the
grid view. The problem is the selector used to apply that styling is too
generic and not specific to the grid view which can alter the position
of the no content helper in other views.

This commit makes sure the style added for the grid view is only applied
inside the grid view and not somewhere else.

task-4207210

closes odoo/enterprise#70818

X-original-commit: c0b592efe498e7ac89b4de570e28394f18e6727d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>