[FIX] l10n_pe_edi_pos: sign invoice before rendering invoice pdf
When invoicing an order from the PoS the invoice was not signed by the
government before rendering the pdf of the invoice. This lead to a
missing QR Code on the invoice.

Steps to reproduce:
-------------------
* Install l10n_pe_edi_pos module
* Open the PoS
* Make an order and invoice it
> Observation: The invoice should contains a QR Code attesting that the
document has been signed by the government but it's not

Why the fix:
------------
The issue was that the invoice was sent to the government after the
invoice was rendered. To fix it we make the call to the government
manually instead of waiting for the CRON to send it.
We also cancel the CRON for this specific invoice so that it is not
called twice. This is based on what is done here : https://github.com/odoo/odoo/blob/647197c0eae7ef786fa3d1aeafeafd1d14fe0fae/addons/l10n_es_pos_tbai/models/pos_order.py#L18-L29

opw-4165399

closes odoo/enterprise#71954

X-original-commit: ea5ca1514adc76aaf254dfaaa91a9ec0bb9514dd
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>