[REF] knowledge: convert the file behavior to an embedded component
The file embedded component was moved to html_editor. This commits extends the
feature to add the Macro capabilities of the file block, allowing to attach it
to a previously visited record, or to send it by email through that previously
visited record.

task-3672747

X-original-commit: 09950acf1bd605faa633d45feddc98e41200c3a6
Part-of: odoo/enterprise#71082
Related: odoo/odoo#182145
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>