[FIX] planning: make resource is set on shift before checking if flexible
Before this commit, when the user copy shifts of the previous weeks and
in that previous week he has an open shift, he will have a traceback will
be raised because we check if the resource of the shift is flexible or
not to be able to correctly compute the allocated hours of the splits
generated with the split to split. The problem is an open shift has no
resource so it is no needed to check if the resource is flexible.

This commit checks if the shift is an open shift or not before checking
if the resource is flexible or not.
This commit also fixes another when we try to copy previous an open
shift planned during 2 days, the allocate hours is computed based on the
company schedule instead of the real time span between the date end and
date start as we do in the `_compute_allocated_hours` method.

task-4207839

X-original-commit: dc7abe43ab4b4b3dd9c659aecfd2f5976d9c0ec8
Part-of: odoo/enterprise#71278
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>