[FIX] documents_project: display document link in project task portal view
Steps to reproduce:
- Install document and project.
- Create a portal user.
- Create the project and the task.
- Upload the document in the task.
- Share the document with the portal user.
- Share the project with read access to the portal user.
- Login with the portal user.
- Open the shared project and created task.

Issue:
'Document' is missing from the left sidebar, even after sharing with portal user

Cause:
'Document' is in the link section of the sidebar. It will be visible when any
task link is shared, eg, invoice, sale order, and helpdesk ticket. So, when
there is only a document shared it doesn't display.

Solution:
Display the 'Document' when the document is shared with the portal user.

task-4194318

closes odoo/enterprise#75408

X-original-commit: 9afbc11815e85e5b0ffcdb5e1776d066105e7703
Related: odoo/odoo#190182
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>