[IMP] mrp_workorder: Produce partial qty
Allow operators to register partial production on
work orders, thereby liberating work orders in the
queue before marking operations as done.
This will make it much more flexible for users and
also reduce the need for creating backorders, which
can be quite messy and difficult to understand.

Task 3504085
odoo/odoo#181362

closes odoo/enterprise#70649

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>