[FIX] pos_restaurant_appointment: fix booking backend view
Before when creating two appointment one by one, an error was raised for
the second appointment. This was due to the fact that the calendar view
was refreshed with a wrong active id.

This commit remove the view refresh since the view itself make a RPC to
get the new data.

closes odoo/enterprise#70782

X-original-commit: 1bd24f27dae48970d46ea2230fe6aa073b68e186
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>