[IMP] account_loans: reversed reclassification optimization
When dealing with loans, it isn't unusual to have it spanned
over many years (20 or 25 years in Belgium for instance).
This means that we have 25 years * 12 months * 3 moves = 900 moves
to create upon confirmation of the loan in Odoo.

The problem here, is that when reversing the reclassification moves,
the `_reverse_moves()` method was doing a reverse of 300 moves one by
one which is very slow.

Instead, we will now create the reverse reclassification moves manually
and simply batch their creation and posting which is much faster.

For the example mentioned, we went from ~4 minutes to 40 seconds.

task-id/opw: none

closes odoo/enterprise#70658

Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>