[FIX] pos_iot: ingenico payment traceback
Starting d1a94147060083da54f9423f8c379676fbb7db66, when making ingenico payment,
the pos app sends an `NaN` as `TransactionID`. As a result, users are unable to
receive payments from Ingenico terminal. This change proposes to take the last
13 characters of the payment's uuid (after removing the `-` separators) and
convert it as integer which will serve as the `TransactionID`. We're only taking
the 13 characters because it's the number of characters that can be converted to
JS int without exponent.

closes odoo/enterprise#70765

X-original-commit: 485cb527805a46ed43692d00f8dadd03c89b8380
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>