[FIX] l10n_fr_reports: show EC Sales List for France
Currently the EC Sales List menuitem is not visible when only installing
`l10n_fr_reports` (and no other localization activating the item).

To reproduce it you need to install `l10n_fr_reports` on a database
where the item is not already visible. (It does not work on the runbot.)

closes odoo/enterprise#71129

X-original-commit: 54a4c99cedb478b9a1f17bfe36f34ae01b3b9938
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>