[IMP] stock_barcode: precise JS test tours
The method `processBarcode` is now slightly slower when processing a
single barcode. In consequence, some not so precise triggers in the
tours run too faster and don't correctly wait the previous trigger's
action to be completed.

This commit changes triggers to make tours more reliable even if they
are slower.

X-original-commit: 2e67f0f94f85fbe4cc2d2031a61d14accd336fbd
Part-of: odoo/enterprise#70835
Signed-off-by: Steve Van Essche <svs@odoo.com>