[FIX] timesheet_grid_holidays: filter tasks by project
Steps to reproduce:
- Install Timesheet when on Time Off
- Timesheet > Start > Input a project in the header
- Try to input a task

Unlike in previous versions, the tasks are not filtered to only show the
belonging to the selected project. This happens because
timesheet_grid_holidays adds a value to the fieldInfo domain, having a
non empty domain means it is picked over field.domain, which contains
the necessary filters otherwised used on tasks. This only happes for
task_id because we change the field type into 'task_with_hours' to apply
a widget specific to this field.

opw-4204658

closes odoo/enterprise#71091

X-original-commit: 738b4e9f201475d3e1615c3df1084032beb2073c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>