[FIX] sale_purchase_stock_inter_company_rules: add test for reservation
In the community-side, we allow the reservation on all transit locations
(i.e. Inter-Company transit as well).
Adds a test where an interco PO is generated in draft then confirmed
after the delivery was done in the other company, as it can now reserve
on the already delivered quantities.
Also adapts another test, now that we can reserve on the Inter-Company
transit location.

Task-4207078

closes odoo/enterprise#71156

X-original-commit: 94812fa4350eaf587536ab0fb6f10d1852b24061
Related: odoo/odoo#182252
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>